# internal/appview/openapi.yaml
openapi: 3.0.3
info:
  title: HashPost AppView API
  description: HashPost AppView provides a unified API for accessing HashPost data aggregated from the PDS
  version: 1.0.0
  contact:
    name: HashPost Team
    email: team@hashpost.com
servers:
  - url: http://localhost:8081
    description: Development server
  - url: https://appview.hashpost.com
    description: Production server
paths:
...
  /api/v1/admin/pds/servers:
    get:
      summary: List PDS servers
      description: Get statistics for all connected PDS servers
      operationId: listPDSServers
      tags:
        - PDS Management
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of PDS servers with statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items:
                      $ref: '#/components/schemas/PDSServerStats'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/pds/{endpoint}:
    get:
      summary: Get PDS server details
      description: Get detailed information about a specific PDS server
      operationId: getPDSServerDetails
      tags:
        - PDS Management
      security:
        - bearerAuth: []
      parameters:
        - name: endpoint
          in: path
          required: true
          schema:
            type: string
          description: PDS server endpoint URL (URL encoded)
      responses:
        '200':
          description: PDS server details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PDSServerDetails'
        '400':
          description: Invalid endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: PDS server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/pds/{endpoint}/users:
    get:
      summary: List users from PDS server
      description: Get users authenticated from a specific PDS server
      operationId: listPDSServerUsers
      tags:
        - PDS Management
      security:
        - bearerAuth: []
      parameters:
        - name: endpoint
          in: path
          required: true
          schema:
            type: string
          description: PDS server endpoint URL (URL encoded)
        - name: limit
          in: query
          description: Maximum number of users to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of users to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of users from PDS server
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithRoles'
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          description: Invalid endpoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: PDS server not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Vote endpoints
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details

    # RBAC Schemas
    UserWithRoles:
      type: object
      required:
        - userDid
        - handle
        - roleCount
      properties:
        userDid:
          type: string
          description: User's DID
        handle:
          type: string
          description: User's handle
        displayName:
          type: string
          description: User's display name
        pdsSource:
          type: string
          description: PDS server endpoint where user data is stored
        isLocal:
          type: boolean
          description: Whether user is stored on HashPost PDS
        lastSeenAt:
          type: string
          format: date-time
          description: Last time user authenticated
        roleCount:
          type: integer
          description: Number of roles assigned to the user

    # PDS Management Schemas
    PDSServerStats:
      type: object
      required:
        - pdsEndpoint
        - userCount
        - activeUsers24h
      properties:
        pdsEndpoint:
          type: string
          description: PDS server endpoint URL
        userCount:
          type: integer
          description: Total number of users from this PDS
        lastActivity:
          type: string
          format: date-time
          description: Last activity timestamp
        activeUsers24h:
          type: integer
          description: Number of users active in last 24 hours
    PDSServerDetails:
      type: object
      required:
        - pdsEndpoint
        - userCount
        - activeUsers24h
        - activeUsers7d
        - activeUsers30d
      properties:
        pdsEndpoint:
          type: string
          description: PDS server endpoint URL
        userCount:
          type: integer
          description: Total number of users from this PDS
        lastActivity:
          type: string
          format: date-time
          description: Last activity timestamp
        firstUserCreated:
          type: string
          format: date-time
          description: When the first user from this PDS was created
        activeUsers24h:
          type: integer
          description: Number of users active in last 24 hours
        activeUsers7d:
          type: integer
          description: Number of users active in last 7 days
        activeUsers30d:
          type: integer
          description: Number of users active in last 30 days
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from PDS authentication
...